+++ common/src/main/java/whocraft/tardis_refined/client/TRShaders.java
@@ -1,16 +1,17 @@
 package whocraft.tardis_refined.client;
 
 import com.google.common.collect.ImmutableList;
+import com.mojang.blaze3d.platform.GlUtil;
 import com.mojang.blaze3d.vertex.VertexFormat;
-import net.minecraft.client.Minecraft;
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
 import net.minecraft.client.renderer.RenderStateShard;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.ShaderInstance;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.phys.Vec3;
 import org.apache.commons.lang3.tuple.Triple;
+import whocraft.tardis_refined.TRConfig;
 
-import java.util.List;
 import java.util.Objects;
 
 import static com.mojang.blaze3d.vertex.DefaultVertexFormat.NEW_ENTITY;
@@ -20,8 +21,21 @@ public class TRShaders {
     public static ShaderInstance GLOW_SHADER;
     public static ShaderInstance SNOW_SHADER;
 
+
+    @Environment(EnvType.CLIENT)
+    public static boolean shouldUseCompatMode() {
+        return !GlUtil.getVendor().toLowerCase().contains("nvidia");
+    }
+
+
     public static RenderType translucentWithSnow(ResourceLocation texture, boolean show) {
         ResourceLocation snowTexture = new ResourceLocation("minecraft", "textures/block/snow.png");
+
+
+        // Has to be separate, using java OR operator would allow the snow to render even if the user has internal shaders turned off
+        if(!TRConfig.CLIENT.USE_INTERNAL_SHADERS.get()){
+            return RenderType.entityTranslucent(texture);
+        }
 
         if (!show) return RenderType.entityTranslucent(texture);