+++ common/src/main/java/whocraft/tardis_refined/constants/ModMessages.java
@@ -46,7 +46,8 @@ public class ModMessages {
     public static final String CONFIG_IP_COMPAT = config("immersive_portals");
     public static final String CONFIG_CONTROL_NAMES = config("control_names");
     public static final String CONFIG_IDLE_CONSOLE_ANIMS = config("console_idle_animations");
-    public static final String DISABLE_FANCY_RENDERING = config("disable_fancy_rendering");
+    public static final String CONFIG_RENDER_VORTEX_IN_DOOR = config("config_render_vortex_in_door");
+    public static final String CONFIG_CUSTOM_SHADERS = config("config_custom_shaders");
     public static final String CONFIG_BANNED_DIMENSIONS = config("banned_dimensions");
     public static final String CMD_DIM_NOT_A_TARDIS = cmdOutput("dim_not_a_tardis");
     public static final String CMD_NO_INTERNAL_DOOR = cmdOutput("no_internal_door");